backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
pytest
httpx
